migrate to using BIgDecimal over Double

+++ chip-in-backend/src/main/java/dev/asif/chipinbackend/dto/ExpenseParticipantDTO.java
@@ -3,11 +3,13 @@ package dev.asif.chipinbackend.dto;
 import dev.asif.chipinbackend.model.ExpenseParticipant;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ExpenseParticipantDTO {
     private Long userId;
-    private double paidAmount;
-    private double shareAmount;
+    private BigDecimal paidAmount;
+    private BigDecimal shareAmount;
 
     public ExpenseParticipantDTO(ExpenseParticipant participant) {
         this.userId = participant.getUser().getId();
