+++ chip-in-backend/src/main/java/dev/asif/chipinbackend/dto/PayerDTO.java
@@ -2,8 +2,10 @@ package dev.asif.chipinbackend.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PayerDTO {
     private Long userId;
-    private double paidAmount;
+    private BigDecimal paidAmount;
 }
